Move the studio agent and the audio probe into their own folders
Ten studio-agent-* files sat at the root of wasmaudioworklet/ alongside the
app's own modules, and this PR was about to add two more.

  wasmaudioworklet/studio-agent/   client, prompt, tools-def, tools-core,
                                   nearai-core, and their tests
  wasmaudioworklet/audioprobe/     audioanalysis, instrumentprobe

The probe deliberately does NOT live under studio-agent/. It is shared with
tools/instrumenttest/render.mjs, and filing a CLI dependency under the agent
would misstate who owns it -- the next thing that wants to measure audio would
have to reach into the agent's folder to get it.

Names lose the redundant prefix (studio-agent/client.js, not
studio-agent/studio-agent-client.js). Git tracked all ten as renames, so the
diff stays reviewable.

Each folder gets a README covering what the files are and, more usefully, WHY
the feedback loops exist -- the agent cannot hear the music, so every tool has
to report what it can actually establish.

Verified: 52 unit tests, the gitproxy/nearai-proxy suites, the CLI harness
still renders drumkit.dsp correctly, and 8 browser e2e tests confirm app.js
still loads the agent from its new path.
