+++ wasmaudioworklet/app.js
@@ -2,7 +2,7 @@ import { } from './audioworkletpolyfill.js';
 import { initAudioWorkletNode } from './audioworkletnode.js';
 import { initVisualizer, setCurrentTimeSeconds as setVisualizerCurrentTimeSeconds } from './visualizer/defaultvisualizer.js';
 import { initEditor } from './editorcontroller.js';
-import { initStudioAgent } from './studio-agent-client.js';
+import { initStudioAgent } from './studio-agent/client.js';
 import { toggleSpinner } from './common/ui/progress-spinner.js';
 import { modalAlert } from './common/ui/modal.js';
 import apphtml from './app.html.js';
