+++ wasmaudioworklet/package.json
@@ -32,7 +32,7 @@
         "test-claude-bridge": "npx playwright test e2e/claude-bridge.spec.js",
         "near-sandbox": "docker run --rm -p 3030:8080 ghcr.io/petersalomonsen/near-git-storage/sandbox:main",
         "test-faust": "node ../tools/faust2as/generate-test-sources.js && npx playwright test e2e/faust2as-compilation.spec.js",
-        "test-agent-tools": "node --test studio-agent-tools-core.test.mjs studio-agent-nearai-core.test.mjs studio-agent-tools-def.test.mjs",
+        "test-agent-tools": "node --test studio-agent/tools-core.test.mjs studio-agent/nearai-core.test.mjs studio-agent/tools-def.test.mjs",
         "test-quickjs-sandbox": "node --test quickjs-sandbox-poc.test.mjs",
         "test-gitproxy": "node --test gitproxy.test.mjs nearai-proxy.test.mjs"
     },
